% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_cfa}
\alias{dat_cfa}
\title{A Sample Dataset Based On a
Confirmatory Factor Analysis Model
(For Testing)}
\format{
A data frame with six variables:
\describe{
\item{x1}{Indicator}
\item{x2}{Indicator}
\item{x3}{Indicator}
\item{x4}{Indicator}
\item{x5}{Indicator}
\item{x6}{Indicator}
}
}
\usage{
dat_cfa
}
\description{
Generated from
a confirmatory factor analysis
model (n = 200).
}
\details{
The model used to generate
this dataset:

\if{html}{\out{<div class="sourceCode">}}\preformatted{f1 =~ x1 + x2 + x3 + x5
f2 =~ x3 + x4 + x5 + x6
f1 ~~ f2
}\if{html}{\out{</div>}}
}
\keyword{datasets}
