% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_path_model}
\alias{dat_path_model}
\title{A sample dataset based on a path model (for testing)}
\format{
A data frame with four variables:
\describe{
\item{x1}{Predictor}
\item{x2}{Predictor}
\item{x3}{Mediator}
\item{x4}{Outcome}
}
}
\usage{
dat_path_model
}
\description{
Generated from the following path model, n = 100
}
\details{
The model used to generate this dataset:

\if{html}{\out{<div class="sourceCode">}}\preformatted{x1 ~~ x2
x3 ~ x1 + x2
x4 ~ x3 + x1 + x2
}\if{html}{\out{</div>}}
}
\keyword{datasets}
