% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_path_model_p06}
\alias{dat_path_model_p06}
\title{A sample dataset based on a complex path model (for testing)}
\format{
A data frame with six variables:
\describe{
\item{x1}{Predictor}
\item{x2}{Predictor}
\item{x3}{Predictor}
\item{y4}{Mediator}
\item{y5}{Mediator}
\item{y6}{Outcome}
}
}
\usage{
dat_path_model_p06
}
\description{
Generated from the following path model, n = 200
}
\details{
The model used to generate this dataset:

\if{html}{\out{<div class="sourceCode">}}\preformatted{y4 ~  x1 + x2 + x3
y5 ~  y4 + x1 + x2
y6 ~  y4 + y5 + x1 + x2 + x3
x1 ~~ x2 + x3
x2 ~~ x3
}\if{html}{\out{</div>}}
}
\keyword{datasets}
