% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_sem}
\alias{dat_sem}
\title{A Sample Dataset Based On a
Structural Model (For Testing)}
\format{
An object of class \code{data.frame} with 250 rows and 16 columns.
}
\usage{
dat_sem
}
\description{
Generated from
a structural model with
latent variables (n = 250).
}
\details{
The model to be fitted:

\if{html}{\out{<div class="sourceCode">}}\preformatted{f1 =~ x1 + x2 + x3 + x4
f2 =~ x5 + x6 + x7 + x8
f3 =~ x9 + x10 + x11 + x12
f4 =~ x13 + x14 + x15 + x16
f3 ~ f1 + f2
f4 ~ f3
}\if{html}{\out{</div>}}
}
\keyword{datasets}
