% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_serial_4_weak}
\alias{dat_serial_4_weak}
\title{A sample dataset based on a serial mediation path model
with weak paths (for
testing)}
\format{
A data frame with four variables:
\describe{
\item{x}{Predictor}
\item{m1}{Mediator}
\item{m2}{Mediator}
\item{y}{Outcome}
}
}
\usage{
dat_serial_4_weak
}
\description{
Generated from the following path model, n = 100
}
\details{
The model used to generate this dataset:

\if{html}{\out{<div class="sourceCode">}}\preformatted{m1 ~ x
m2 ~ m1 + x
y  ~ m2 + m1 + x
}\if{html}{\out{</div>}}
}
\keyword{datasets}
