% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drop.R
\name{get_drop}
\alias{get_drop}
\title{Models That Are More Restricted}
\usage{
get_drop(
  sem_out,
  must_drop = NULL,
  must_not_drop = NULL,
  df_change = 1,
  model_id = NA,
  keep_correct_df_change = TRUE
)
}
\arguments{
\item{sem_out}{The output from an
SEM function. Currently support
\link[lavaan:lavaan-class]{lavaan::lavaan} only.
Usually the one used \code{\link[=get_add]{get_add()}} or
\code{\link[=get_drop]{get_drop()}} to generate the
list of models.}

\item{must_drop}{A character vector
of parameters, named in
\code{lavaan::lavaan()} style (e.g.,
\code{"y ~ x"}), that must be included.
Default is \code{NULL}.}

\item{must_not_drop}{A character
vector of parameters, named in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} style (e.g.,
\code{"x1 ~~ x1"}), that must not be
included. Default is \code{NULL}.}

\item{df_change}{How many degrees
of freedom away in the list. All
models with \emph{df} change less than
or equal to this number will be
included. Default is 1.}

\item{model_id}{The identification
number of the starting model.
Default is \code{NA}, no identification
number.}

\item{keep_correct_df_change}{Keep
only tables with actual \emph{df} change
equal to expected \emph{df} change.}
}
\value{
A named list of parameter
tables to be used by
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} or \code{\link[=update]{update()}}
for fitting a model with the added
parameters.
}
\description{
Generate the list of
models with one less free parameter
or constraint (and one more
degree of freedom).
}
\details{
Generate the list of models
with one less free parameter
or constraint (and one more degree
of freedom).
}
\examples{
# To Do

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
