% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelbpp-package.R
\docType{package}
\name{modelbpp-package}
\alias{modelbpp}
\alias{modelbpp-package}
\title{modelbpp: Model BIC Posterior Probability}
\description{
Fits the neighboring models of a fitted structural equation model and assesses the model uncertainty of the fitted model based on BIC posterior probabilities, using the method presented in Wu, Cheung, and Leung (2020) \doi{10.1080/00273171.2019.1574546}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/modelbpp/}
  \item Report bugs at \url{https://github.com/sfcheung/modelbpp/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Huiping Wu
  \item Shing On Leung (\href{https://orcid.org/0000-0002-0962-339X}{ORCID})
}

}
\keyword{internal}
