% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.partables.R
\name{print.partables}
\alias{print.partables}
\title{Print a \code{partables}-Class Object}
\usage{
\method{print}{partables}(x, max_tables = 10, ...)
}
\arguments{
\item{x}{A \code{partables}-class object.}

\item{max_tables}{The maximum number
of modifications to be printed.
Default is 10.}

\item{...}{Optional arguments.
Ignored.}
}
\value{
\code{x} is returned invisibly.
Called for its side effect.
}
\description{
Print the content of
an \code{partables}-class object
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

x <- 123
print(x)

}
\seealso{
A \code{partables}-class object
is generated by \code{get_add()} and
\code{get_drop()}.
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
