% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sem_outs.R
\name{print.sem_outs}
\alias{print.sem_outs}
\title{Print an \code{sem_outs}-Class Object}
\usage{
\method{print}{sem_outs}(x, max_models = 10, ...)
}
\arguments{
\item{x}{An \code{sem_outs}-class object.}

\item{max_models}{The maximum number
of models to be printed. Default
is 10.}

\item{...}{Optional arguments.
Ignored.}
}
\value{
\code{x} is returned invisibly.
Called for its side effect.
}
\description{
Print the content of
an \code{sem_outs}-class object
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{

x <- 123
print(x)

}
\seealso{
An \code{sem_outs}-class object
is generated by \code{fit_many()}.
}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
