% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_add.R
\name{get_add}
\alias{get_add}
\title{Models That Are Less Restricted}
\usage{
get_add(
  sem_out,
  must_add = NULL,
  must_not_add = NULL,
  remove_constraints = TRUE,
  exclude_error_cov = TRUE,
  df_change = 1,
  model_id = NA,
  keep_correct_df_change = TRUE
)
}
\arguments{
\item{sem_out}{The output from an
structural equation modeling
function. Currently support
\link[lavaan:lavaan-class]{lavaan::lavaan} only. Usually
the one used in \code{\link[=get_add]{get_add()}} or
\code{\link[=get_drop]{get_drop()}} to generate the list of
models.}

\item{must_add}{A character vector
of parameters, named in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} style (e.g.,
\code{"y ~ x"}), that must be added.
Default is `NULL``.}

\item{must_not_add}{A character
vector of parameters, named in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} style (e.g.,
\code{"x1 ~~ x1"}), that must not be
added. Default is \code{NULL}.}

\item{remove_constraints}{Whether
equality constraints will be
removed. Default is ``TRUE`.}

\item{exclude_error_cov}{Exclude
error covariances of indicators.
Default is \code{TRUE}.}

\item{df_change}{How many degrees
of freedom (\emph{df}) away in the list.
All models with \emph{df} change less than
or equal to this number will be
included. Default is 1.}

\item{model_id}{The identification
number of the starting model.
Default is \code{NA}, no identification
number.}

\item{keep_correct_df_change}{Keep
only tables with actual \emph{df} change
equal to expected \emph{df} change.}
}
\value{
An object of the class
\code{partables}, a named list of parameter
tables, each of them to be used by
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} or \code{\link[=update]{update()}}
for fitting a model with the added
parameters.
}
\description{
Generate the list of
models with one more free
parameter (and one less degree of
freedom).
}
\details{
Generate the list of models
with one more free parameter (and one
less degree of freedom, \emph{df}).
}
\examples{
# To Do

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
