% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_set.R
\name{model_set}
\alias{model_set}
\title{Assessment of a Set of
Neighboring Models}
\usage{
model_set(
  sem_out,
  must_add = NULL,
  must_not_add = NULL,
  must_drop = NULL,
  must_not_drop = NULL,
  remove_constraints = TRUE,
  exclude_error_cov = TRUE,
  df_change_add = 1,
  df_change_drop = 1
)
}
\arguments{
\item{sem_out}{The output from an
SEM function. Currently support
\link[lavaan:lavaan-class]{lavaan::lavaan} only.
Usually the one used \code{\link[=get_add]{get_add()}} or
\code{\link[=get_drop]{get_drop()}} to generate the
list of models.}

\item{must_add}{A character vector
of parameters, named in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} style (e.g.,
\code{"y ~ x"}), that must be added.
Default is `NULL``.}

\item{must_not_add}{A character
vector of parameters, named in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} style (e.g.,
\code{"x1 ~~ x1"}), that must not be
added. Default is \code{NULL}.}

\item{must_drop}{A character vector
of parameters, named in
\code{lavaan::lavaan()} style (e.g.,
\code{"y ~ x"}), that must be included.
Default is \code{NULL}.}

\item{must_not_drop}{A character
vector of parameters, named in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} style (e.g.,
\code{"x1 ~~ x1"}), that must not be
included. Default is \code{NULL}.}

\item{remove_constraints}{Whether
equality constraints will be
removed. Default is ``TRUE`.}

\item{exclude_error_cov}{Exclude
error covariances of indicators.
Default is \code{TRUE}.}

\item{df_change_add}{How many degrees
of freedom (\emph{df}) away in the list
when adding parameters or releasing
constraints. All models with \emph{df}
change less than or equal to this
number will be included. Default is
1.}

\item{df_change_drop}{How many \emph{df}
away in the list when removing
parameters. All models with \emph{df}
change less than or equal to this
number will be included. Default is
1.}
}
\value{
A list with the following
elements:
\itemize{
\item \code{model}: A named list of parameter
tables. Each represent the models
identified.
\item \code{bic}: A numeric vector, of the
same length as \code{model}. The BIC of
each model.
\item \code{postprob}: A numeric vector, of
the same length as \code{model}. The
posterior probability of each
model.
\item \code{fit}: A named list of
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} output objects
or \code{\link[=update]{update()}} for fitting a model
with the added parameters, of the
same length as \code{model}.
\item \code{change}: A numeric vector, of the
same length as \code{model}. The change
in model df for each fit. A
positive number denotes one less
free parameter. A negative number
denotes one more free parameter or
one less constraint.
\item \code{converged}: A named vector of
boolean values, of the same length
as \code{model}. Indicates whether each
fit converged or not.
\item \code{post_check}: A named vector of
boolean values, of the same length
as \code{model}. Indicates whether the
solution of each fit is admissible
or not. Checked by
\code{\link[lavaan:lavInspect]{lavaan::lavInspect()}}.
}
}
\description{
Identify neighboring
models, fit them, and return the
posterior probabilities.
}
\details{
(TO-DO)
}
\examples{
# To Do

}
\author{
Shu Fai Cheung \url{https://orcid.org/0000-0002-9871-9448}
}
